+++ src/components.d.ts
@@ -885,7 +885,19 @@ export namespace Components {
         "value": number;
     }
     /**
-     * The progress-radial component is a circular progress bar that indicates the current completion of a task.
+     * The progress-radial component is a circular progress bar that indicates the current completion of a task. 
+     * ## Accessibility guidelines 💡
+     * > Aria attributes and how to display the progress-radial depend on the use case in your application :
+     * >
+     * > - **Case 1 : decorative**
+     * > If the progress-radial is used as a decoration _(if removed, the user doesn't lose any relevant information)_ or in the
+     * > context of another component _(such as progress-radial in a card)_ => **you don't need to show the label nor add an aria-label**.
+     * >
+     * > - **Case 2 : informative**
+     * > If the progress-radial is used to convey important information _(e.g., form completion status, dashboard KPI)_, you need to :
+     * >   - **Provide a visible label** that describes the purpose of the progress-radial.
+     * >   - **Set the `showLabel` property to `true`** to show the percentage inside the progress-radial.
+     * >   - Optionally, use aria-label to provide an accessible name if a visible label is not present.
      */
     interface WcsProgressRadial {
         /**
@@ -2046,7 +2058,19 @@ declare global {
         new (): HTMLWcsProgressBarElement;
     };
     /**
-     * The progress-radial component is a circular progress bar that indicates the current completion of a task.
+     * The progress-radial component is a circular progress bar that indicates the current completion of a task. 
+     * ## Accessibility guidelines 💡
+     * > Aria attributes and how to display the progress-radial depend on the use case in your application :
+     * >
+     * > - **Case 1 : decorative**
+     * > If the progress-radial is used as a decoration _(if removed, the user doesn't lose any relevant information)_ or in the
+     * > context of another component _(such as progress-radial in a card)_ => **you don't need to show the label nor add an aria-label**.
+     * >
+     * > - **Case 2 : informative**
+     * > If the progress-radial is used to convey important information _(e.g., form completion status, dashboard KPI)_, you need to :
+     * >   - **Provide a visible label** that describes the purpose of the progress-radial.
+     * >   - **Set the `showLabel` property to `true`** to show the percentage inside the progress-radial.
+     * >   - Optionally, use aria-label to provide an accessible name if a visible label is not present.
      */
     interface HTMLWcsProgressRadialElement extends Components.WcsProgressRadial, HTMLStencilElement {
     }
@@ -3225,7 +3249,19 @@ declare namespace LocalJSX {
         "value"?: number;
     }
     /**
-     * The progress-radial component is a circular progress bar that indicates the current completion of a task.
+     * The progress-radial component is a circular progress bar that indicates the current completion of a task. 
+     * ## Accessibility guidelines 💡
+     * > Aria attributes and how to display the progress-radial depend on the use case in your application :
+     * >
+     * > - **Case 1 : decorative**
+     * > If the progress-radial is used as a decoration _(if removed, the user doesn't lose any relevant information)_ or in the
+     * > context of another component _(such as progress-radial in a card)_ => **you don't need to show the label nor add an aria-label**.
+     * >
+     * > - **Case 2 : informative**
+     * > If the progress-radial is used to convey important information _(e.g., form completion status, dashboard KPI)_, you need to :
+     * >   - **Provide a visible label** that describes the purpose of the progress-radial.
+     * >   - **Set the `showLabel` property to `true`** to show the percentage inside the progress-radial.
+     * >   - Optionally, use aria-label to provide an accessible name if a visible label is not present.
      */
     interface WcsProgressRadial {
         /**
@@ -3986,7 +4022,19 @@ declare module "@stencil/core" {
              */
             "wcs-progress-bar": LocalJSX.WcsProgressBar & JSXBase.HTMLAttributes<HTMLWcsProgressBarElement>;
             /**
-             * The progress-radial component is a circular progress bar that indicates the current completion of a task.
+             * The progress-radial component is a circular progress bar that indicates the current completion of a task. 
+             * ## Accessibility guidelines 💡
+             * > Aria attributes and how to display the progress-radial depend on the use case in your application :
+             * >
+             * > - **Case 1 : decorative**
+             * > If the progress-radial is used as a decoration _(if removed, the user doesn't lose any relevant information)_ or in the
+             * > context of another component _(such as progress-radial in a card)_ => **you don't need to show the label nor add an aria-label**.
+             * >
+             * > - **Case 2 : informative**
+             * > If the progress-radial is used to convey important information _(e.g., form completion status, dashboard KPI)_, you need to :
+             * >   - **Provide a visible label** that describes the purpose of the progress-radial.
+             * >   - **Set the `showLabel` property to `true`** to show the percentage inside the progress-radial.
+             * >   - Optionally, use aria-label to provide an accessible name if a visible label is not present.
              */
             "wcs-progress-radial": LocalJSX.WcsProgressRadial & JSXBase.HTMLAttributes<HTMLWcsProgressRadialElement>;
             "wcs-radio": LocalJSX.WcsRadio & JSXBase.HTMLAttributes<HTMLWcsRadioElement>;